Update README.md
Smart Dustbin project based on IoT

components:
#Hardware:
*ESP8266 (1)
*Servomotor (2)
*Ultrasonic Sensor (3)
*Moisture sensor (1)
*I2C Module
*Bin
*Lid
*Cardboard
*Jumper-wires

#Software:
*Arduino IDE
#Blynk IoT

Specifications of this Smart Dustbin is:
*Automatic lid opening/closing
*Waste Monitoring
*Waste Segregating

Working:
When a person comes closer to the bin a Ultrasonic sensor senses the distance and if the person is very close to bin then the Servomotor lifts up the lid and when the person moves away automatically the lid closes. When a person keeps the waste into the bin and the Moisture sensor senses the wetness the Servomotor rotates to left by this the waste falls left side and if the waste is kept into bin and if the Moisture sensor senses no wetness then the Servomotor turns to right and the waste falls on right side. Their are two Ultrasonic for monitoring the level of waste one on right side another on left side, if any one of the side gets full by waste, then a message will be sent to the owner.